heuristic for song length determination
song length calculation of lorge.fur now takes 18ms (from ~1 hour)

+++ src/engine/song.h
@@ -171,7 +171,7 @@ struct DivGroovePattern {
 struct DivSongTimestamps {
   // song duration (in seconds and microseconds)
   TimeMicros totalTime;
-  int totalTicks;
+  uint64_t totalTicks;
   int totalRows;
 
   // loop region (order/row positions)
